+++ discord_twitter_webhooks/send_webhook.py
@@ -20,13 +20,12 @@ def send_embed_webhook(
     """Send embed to Discord webhook.
 
     Args:
-        avatar (str): Avatar URL
-        tweet_id (int): Tweet id
-        link_list (list[str]): List of links from the tweet
-        text (str): Text from the tweet
-        webhook (str, optional): Webhook URL. Defaults to environment
-        variable WEBHOOK_URL.
-        twitter_card_image (str, optional): Twitter meta image.
+        avatar: Avatar URL
+        tweet_id: Tweet id
+        link_list: List of links from the tweet
+        text: Text from the tweet
+        webhook: Webhook URL. Defaults to environment variable WEBHOOK_URL.
+        twitter_card_image: Twitter meta image.
     """
     settings.logger.debug(f"Tweet: {text}")
 
@@ -45,14 +44,14 @@ def send_embed_webhook(
         elif len(link_list) > 1:
             # Send images to twitter-image-collage-maker
             # (e.g https://twitter.lovinator.space/) and get a collage back.
-            response = requests.get(url=settings.collage_maker_url, params={"tweet_id": tweet_id})  # noqa: E501, pylint: disable=line-too-long
+            response = requests.get(url=settings.collage_maker_url, params={"tweet_id": tweet_id})
 
             if response.status_code == 200:
                 json_data = json.loads(response.text)
-                settings.logger.debug(f"JSON data from twitter-image-collage-maker: {json_data}")  # noqa: E501, pylint: disable=line-too-long
+                settings.logger.debug(f"JSON data from twitter-image-collage-maker: {json_data}")
                 embed.set_image(url=json_data["url"])
             else:
-                settings.logger.error(f"Got {response.status_code} from {settings.collage_maker_url}")  # noqa: E501, pylint: disable=line-too-long
+                settings.logger.error(f"Got {response.status_code} from {settings.collage_maker_url}")
                 embed.set_image(url=link_list[0])
 
     settings.logger.debug(f"Avatar URL: {avatar_url}")
@@ -68,7 +67,7 @@ def send_embed_webhook(
 
     response = hook.execute()
 
-    settings.logger.info(f"Webhook posted for tweet https://twitter.com/i/web/status/{tweet_id}")  # noqa: E501, pylint: disable=line-too-long
+    settings.logger.info(f"Webhook posted for tweet https://twitter.com/i/web/status/{tweet_id}")
     settings.logger.debug(f"Webhook response: {response}")
 
 
@@ -76,9 +75,8 @@ def send_normal_webhook(msg: str, webhook: str = settings.webhook_url):
     """Send normal message to Discord webhook.
 
     Args:
-        msg (str): Message to send
-        webhook (str, optional): Webhook URL. Defaults to environment
-        variable WEBHOOK_URL.
+        msg: Message to send
+        webhook: Webhook URL. Defaults to environment variable WEBHOOK_URL.
     """
     settings.logger.debug(f"Message: {msg}")