+++ tests/test_get.py
@@ -3,7 +3,7 @@ from discord_twitter_webhooks.get import meta_image
 
 class TestGet:
     hello_txt = "Hello @TheLovinator1 #Hello /u/test /r/aww"
-    hello2_txt = "/r/hello r/hello hello/r/hello /u/hello u/hello hello/u/hello"  # noqa: E501, pylint: disable=line-too-long
+    hello2_txt = "/r/hello r/hello hello/r/hello /u/hello u/hello hello/u/hello"
 
     short = "Hello I am short Sadge"