+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to discord-twitter-webhooks will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## Changed
+
+- Removed type hints from docstrings.
+- Line length is now restricted to 120 characters.
+- Remove #noqa comments from code.
+
 ## [2.0.0] - 2022-05-23
 
 ## Changed