+++ CHANGELOG.md
@@ -5,9 +5,13 @@ All notable changes to discord-twitter-webhooks will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.1.0] - 2022-08-09
 
-## Changed
+### Added
+
+- You can now send different rules to different webhooks.
+
+### Changed
 
 - Removed type hints from docstrings.
 - Line length is now restricted to 120 characters.
@@ -15,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [2.0.0] - 2022-05-23
 
-## Changed
+### Changed
 
 - You need to use BEARER_TOKEN to authenticate instead of CONSUMER_KEY,
   CONSUMER_SECRET, ACCESS_TOKEN and ACCESS_TOKEN_SECRET